some changes

+++ app/Http/Controllers/ReportController.php
@@ -2,17 +2,15 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Http\Requests\ReportsRequest;
 use App\Report;
-use App\User;
 use DateTime;
 
 class ReportController extends Controller {
 
     public $puginationReports = 5;
-	public $puginationReportComments = 10;
-	public $lastReports = 5;
+    public $puginationReportComments = 10;
+    public $lastReports = 5;
     public $puginationAdminReports = 15;
 
     public function index() {
@@ -28,7 +26,7 @@ class ReportController extends Controller {
 
     public function show($id) {
         $report = Report::select()->where('id', $id)->first();
-		$report->setRelation('comments', $report->comments()->paginate($this->puginationReportComments));
+        $report->setRelation('comments', $report->comments()->paginate($this->puginationReportComments));
         $lastReports = Report::orderBy('id', 'desc')->take($this->lastReports)->get();
         return view('reports.report', compact('report', 'lastReports'));
     }
@@ -43,23 +41,22 @@ class ReportController extends Controller {
     }
 
     public function store(ReportsRequest $request) {
-        if ($request->method() == 'POST') {
-            $data = $request->all();
-            unset($data['__token']);
-            $date = new DateTime();
-            $data['date'] = $date->format('Y-m-d');
-            if ($request->hasFile('pay_check')) {
-                $data['pay_check'] = $this->addPayCheck($request);
-            };
-            $create = Report::create($data);
-            $id = $create->id;
-            return redirect()->route('adminReports');
-        }
-        return view('home');
+        $data = $request->all();
+        unset($data['__token']);
+        $date = new DateTime();
+        $data['date'] = $date->format('Y-m-d');
+        if ($request->hasFile('pay_check')) {
+            $data['pay_check'] = $this->addPayCheck($request);
+        };
+        $create = Report::create($data);
+        return redirect()->route('adminReports');
     }
 
     public function destroy($id) {
-        $report = Report::find($id);
+        if (!is_numeric($id)) {
+            return false;
+        }
+        $report = Report::findOrFail($id);
         Report::find($id)->comments()->forceDelete();
         $img = $report->pay_check;
         if (is_file(public_path() . '/images/reports/' . $img)) {
@@ -76,16 +73,14 @@ class ReportController extends Controller {
     }
 
     public function update(ReportsRequest $request, $id) {
-        if ($request->method() == "POST") {
-            $data = $request->all();
-            if ($request->hasFile('pay_check')) {
-                $data['pay_check'] = $this->addPayCheck($request);
-            };
-            $editOne = Report::find($id);
-            $editOne->fill($data);
-            $editOne->save();
-            return redirect()->route('adminReports');
-        }
+        $data = $request->all();
+        if ($request->hasFile('pay_check')) {
+            $data['pay_check'] = $this->addPayCheck($request);
+        };
+        $editOne = Report::find($id);
+        $editOne->fill($data);
+        $editOne->save();
+        return redirect()->route('adminReports');
     }
 
     public function addPayCheck($request) {
